finish updating current location on patients' app

+++ backend/src/main/java/ifn372/sevencolors/backend/entities/Location.java
@@ -4,27 +4,28 @@ package ifn372.sevencolors.backend.entities;
  * Created by lua on 15/08/2015.
  */
 public class Location {
-    public float getLat() {
+    public double getLat() {
         return lat;
     }
 
-    public void setLat(float lat) {
+    public void setLat(double lat) {
         this.lat = lat;
     }
 
-    public float getLon() {
+    public double getLon() {
         return lon;
     }
 
-    public void setLon(float lon) {
+    public void setLon(double lon) {
         this.lon = lon;
     }
 
-    private float lat;
-    private float lon;
+    private double lat;
+    private double lon;
 
     public Location(){}
-    public Location(float lat, float lon) {
+
+    public Location(double lat, double lon) {
         this.lat = lat;
         this.lon = lon;
     }
